Move dendritic-unflake to the very bottom.
Looks like more people are noticing the Dendritic pattern now, and it seems (I've read at least two people mention) that the very first example being my non-flakes repository causes more questions and wrong expectations about it being the *Flagship* example. 

So I've moved the link to dendritic-unflake to the very end, made our dendritic repos highligted as real-world implementations where people can explore the benefits of the pattern. 

I've also added Michael Belsanti repo which is awesome and very well structured. 


Please please merge, it is an issue for me that people expect my non-flakes exploration repo to be the flagship example when it is not.